+++ TabletopTweaks/NewComponents/AddAdditionalWeaponDamage.cs
@@ -19,6 +19,9 @@ namespace TabletopTweaks.NewComponents {
             if (CheckWeaponRangeType && !RangeType.IsSuitableWeapon(evt.Weapon)) {
                 return;
             }
+            if (CheckWeaponCatergoy && evt.Weapon.Blueprint.Category != Category) {
+                return;
+            }
             DamageDescription Damage = new DamageDescription {
                 TypeDescription = DamageType,
                 Dice = new DiceFormula(Value.DiceCountValue.Calculate(base.Context), Value.DiceType),
@@ -34,5 +37,7 @@ namespace TabletopTweaks.NewComponents {
         public ContextDiceValue Value;
         public bool CheckWeaponRangeType;
         public WeaponRangeType RangeType;
+        public bool CheckWeaponCatergoy;
+        public WeaponCategory Category;
     }
 }