+++ TabletopTweaks/NewContent/Feats/IntensifiedSpell.cs
@@ -26,7 +26,7 @@ namespace TabletopTweaks.NewContent.Feats {
             var IntensifiedSpellFeat = Helpers.CreateBlueprint<BlueprintFeature>("IntensifiedSpellFeat", bp => {
                 bp.SetName("Metamagic (Intensified Spell)");
                 bp.SetDescription("Your spells can go beyond several normal limitations.\n" +
-                    "An intensified spell increases the maximum number of damage dice by 5 levels. " +
+                    "Benefit: An intensified spell increases the maximum number of damage dice by 5 levels. " +
                     "You must actually have sufficient caster levels to surpass the maximum in order " +
                     "to benefit from this feat. No other variables of the spell are affected, and spells " +
                     "that inflict damage that is not modified by caster level are not affected by this feat.\n" +
@@ -66,7 +66,7 @@ namespace TabletopTweaks.NewContent.Feats {
                 name: "Intensified",
                 icon: Icon_IntensifiedSpellMetamagic,
                 defaultCost: 1,
-                CustomMechanicsFeature.FavoriteMetamagicIntensified
+                favoriteMetamagic: CustomMechanicsFeature.FavoriteMetamagicIntensified
             );
             UpdateSpells();
             FeatTools.AddAsFeat(IntensifiedSpellFeat);